Client_Acquisition_System/requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
pyyaml>=6.0
openai>=1.30.0
apify-client>=1.6.0
outscraper>=5.0.0
pyhunter>=1.8
pandas>=2.0.0
beautifulsoup4>=4.12.0
tldextract>=5.1.0
tenacity>=8.2.0
